+++ src/Button/Button.test.tsx
@@ -1,39 +1,8 @@
 import React from "react";
 import faker from "faker";
-import { render, fireEvent } from "@testing-library/react";
+import { fireEvent } from "@testing-library/react";
 import { Button } from "./Button";
-import { ContentProvider } from "../Content/Content";
-
-const TestButton = ({
-  contentNodes = {},
-  theme = {},
-  lang = "en",
-  contentKey,
-  ...rest
-}) => {
-  const defaultContentNodes = {};
-
-  const defaultTheme = {
-    colors: {
-      primary: {
-        light: "#d0dbfd",
-        dark: "#230cc2"
-      }
-    }
-  };
-
-  return (
-    <ContentProvider
-      content={{
-        theme: { ...defaultTheme, ...theme },
-        contentNodes: { ...defaultContentNodes, ...contentNodes }
-      }}
-      lang={lang}
-    >
-      <Button contentKey={contentKey} {...rest} />
-    </ContentProvider>
-  );
-};
+import { renderWithContent } from "../testUtils";
 
 describe("content", () => {
   test("renders English copy", () => {
@@ -46,9 +15,11 @@ describe("content", () => {
         }
       }
     };
-    const { container } = render(
-      <TestButton lang="en" contentNodes={contentNodes} contentKey="button" />
-    );
+
+    const { container } = renderWithContent({
+      contentNodes,
+      lang: "en"
+    })(<Button contentKey="button" />);
 
     expect(container).toMatchSnapshot();
   });
@@ -63,9 +34,11 @@ describe("content", () => {
         }
       }
     };
-    const { container } = render(
-      <TestButton lang="fr" contentNodes={contentNodes} contentKey="button" />
-    );
+
+    const { container } = renderWithContent({
+      contentNodes,
+      lang: "fr"
+    })(<Button contentKey="button" />);
 
     expect(container).toMatchSnapshot();
   });
@@ -74,12 +47,9 @@ describe("content", () => {
 describe("events", () => {
   test("calls onClick callback when click event is fired", async () => {
     const handleClick = jest.fn(() => {});
-    const { getByTestId } = render(
-      <TestButton
-        data-testid="button"
-        onClick={handleClick}
-        contentKey="hello"
-      />
+
+    const { getByTestId } = renderWithContent({})(
+      <Button data-testid="button" onClick={handleClick} contentKey="hello" />
     );
 
     fireEvent.click(getByTestId("button"));
